+++ src/video_output/ycbcr2rgb/ycbcr_infile.txt
@@ -1,20 +1,30 @@
-0C 70 45
-7B E5 77
-05 16 8F
-47 72 42
-DB 8D 9D
-DA B2 2F
-5C CD CA
-F6 18 6D
-57 D9 12
-2D 02 4E
-BA 94 08
-B1 43 4D
-F5 30 1A
-F1 01 C6
-55 54 2D
-50 F3 A0
-37 67 83
-5B BC AD
-7E 87 63
-70 41 BE
+73 7C 8C
+F6 D0 FF
+3A D8 35
+77 54 D7
+7D A8 97
+ED 04 24
+8C FA E7
+39 7B 61
+42 1C D6
+2B B1 F3
+27 61 F3
+54 07 20
+98 52 58
+24 3B 4B
+AA 02 5C
+A6 26 A2
+B5 63 9A
+92 5C E2
+1C 1D 32
+79 B9 0C
+41 23 E7
+5C 1A C8
+24 6D ED
+C7 0C CF
+4A 3A 64
+09 B3 24
+96 9F 75
+75 C2 85
+5E 33 29
+F6 F6 BD
