+++ src/video_output/ycbcr2rgb/ycbcr_vhdl_outfile.txt
@@ -1,0 +1,30 @@
+836C6B
+FF81FF
+0052D6
+F14928
+9D5FC4
+6BFF10
+FF19FF
+0D5130
+BA2800
+CC0082
+C80000
+00C200
+5FC546
+006200
+77EF00
+D5AD06
+D9AD81
+FF5952
+007600
+00B8DE
+D11800
+C04C00
+BC0002
+FFB700
+227600
+003963
+8694CD
+7C5BEA
+00B700
+FFA3FF
